F.E - Tela Inicial e Tela de Cálculo de Te

+++ AppCalculaMotor/AppCalculaMotor/AppCalculaMotor/Modelo/Motor.cs
@@ -39,7 +39,7 @@ namespace AppCalculaMotor.Modelo
             }
         }
 
-        public Motor(double cilindrada, int numerocilindros, int taxaestatica)
+        public Motor(double cilindrada, int numerocilindros, double taxaestatica)
         {
             double cilindradaIndividual = cilindrada / numerocilindros;
             VolumeCamara = CalcularVolumeCamara(cilindradaIndividual, taxaestatica);
